`kernel`,`kernel_x86_64`: remove common `PerCpu` abstraction
We're not sure how context-switching will work for RISC-V yet, but hemming
it into requiring a bunch of per-CPU abstractions in the common kernel
restricts how RV can handle context switches in the future.

Instead, we push the context-switching machinery wholesale into the
platform, and just call into it from the scheduler. For x64, this
simply pushes the per-CPU access up a level, but constrains it to the
`kernel_x86_64` crate. We also provide a cleaner interface to the
scheduler to the common syscall code by simply... passing it to it as a
parameter.