`kernel`: rewrite scheduler to bring per-CPU state within it
For various reasons, it seems more sensible to have the scheduler hold all
per-CPU state inside the common layer. This was primarily because we're
now having to mess around with locked state for RiscV, but
this seems like a sensible architectural decision for the long-term anyways.

This involves changes to `kernel_x86_64` too - we've brought the scheduler
out of the per-CPU info so this actually makes things simpler.