Context switching on RISC-V!
This required quite a bit of re-organisation across the common kernel and
the RV and x64 kernels - context switching is a tonne nicer on RV than x64,
and after some experimentation works best if you simply store a single
static context for each task, and read and write to it when switching. We
still need to do faffing about on the stack for x64 (or do we? Maybe worth
investigating since we have that weird bug anyways) so we basically have
to give the architecture-layer of the kernel a choice of how it does
context switches.