`seed_riscv`: allocate memory for BARs for PCI devices
On RISC-V, we are responsible for allocating BARs to PCI devices from
suitable regions of the physical address space. This reads the supported
ranges out of the device tree, and allocates them to BARs as PCI devices
are enumerated using an extremely simple bump allocator. It's also responsible
for configuring the PCI devices to respond to memory accesses once this
is done.

This required extra functionality from `pci_types`, which will be landed
upstream in the future.