`lib/poplar`: make `Handle`s `Clone+Copy`
Not sure re tradeoff between the idea that handles were "linear" types
that sort of kept track of objects, and the idea that they shouldn't be a
total pain in the ass to pass around in the raw syscall layer. However,
now we've got the abstracted types in-progress (e.g. a `MemoryObject` type),
the potential for getting confused with handles lasting longer than they should
is less because normal use should be via the higher-level types.