`usb_bus_ehci`: start initializing controller and querying ports
There's quite a lot to USB, even for an older controller such as EHCI!
This makes a start in initializing the controller and querying the ports
for connected devices. After this, we'll actually have to work out how to
send transfers to devices to query their descriptors and such, but so far
it looks like what we have done is working (at least on QEMU)!