`lib/bipqueue`: write a basic version of a bi-partite buffer lockless queue
Attributions are included in the source. I think this is correct, but
lockless algorithms are very new for me so we'll see...

The first use of this - getting bytes from the UART interrupt handler to
the rest of the kernel, really doesn't need the characteristics of a
`BipQueue` (any old ring buffer thing would've done) but I think these
are cool and will be useful in the future for other things, so we might as
well use one?