`poplar`: create `DmaPool` abstraction in new `poplar::ddk` module
This creates a Driver Development Kit (DDK) module and associated feature
in the `poplar` library that will hold userspace abstractions that are only/
mostly useful for drivers. This pulls in `linked_list_allocator`, and probably
a bunch of other stuff in the future, so it's useful to have it feature-gated.

The first abstraction in the DDK is a little `DmaPool` and `DmaObject` etc.
system for dynamically managing memory that needs to be addressed physically.
This has already proven useful in our EHCI driver.

The `ddk` module can be re-exported from our custom `std` too for easier
use.