`lib/usb`: fix the representation of setup packet's `RequestType`s
Wow this took a long time to track down... `mycelium_bitfield` doesn't
work quite how we expected it to, and it took a while because we haven't
set any of these top bits yet. Turns out when a setup packet doesn't make
sense, QEMU just kind of grinds the whole queue to a halt...

Oh well, fixed now.