`platform_bus`: ask device drivers if they can support devices before handing them off
This is preparing for a USB HID driver, which needs to parse a USB device's
configuration hierachy to know if it is a HID device. That's not really
possible to write as a filter we can apply in `platform_bus`, so this is
an extra step where we ask a driver if it's sure it can drive a device
before handing it off. This allows a driver to use a more permissive initial
filter and then refuse devices that don't match more stringent criteria later.

The support for this in `platform_bus` is a bit crap tbh, and will fall over
if drivers actually refuse devices. That won't happen for our initial plans
and can effectively be pushed down the line until we have user async runtimes,
which should allow us to write much more expressive message-handling logic
(hopefully; obviously we don't really know what the ergonomics will look like
yet).

This also really needs docs with nice diagrams and stuff, otherwise it's
probably pretty impenitrable. Work on that at some point.