`usb_bus_ehci`: respect correct(?) ordering of requests + get config hierachy
This gets all the ducks in a row re the hopefully correct ordering of `GetDescriptor`
requests, setting the address, and getting the configuration hierachy.
We won't know how close this is to working until we get a real device running
it, but it seems to work on QEMU at least. Also restructures the transaction
code to take ownership of the `SetupPacket` memory - more work will be needed
to take ownership of the data buffer too and then give it back.