`usb_hid`: add start of very basic USB HID driver
This is the start of a driver for USB HID-class devices. It can detect devices
it can potentially support from the Platform Bus, and then interrogate
their configuration hierachies for HID-class interfaces. Also includes some
definitions for HID descriptors for the future.

Next steps will be actually parsing the HID descriptors (including the Report
descriptors, which look potentially complex), and then working on Interrupt
endpoints to be able to get data from a HID device! I think it would be worth
getting some of the other parts of this infrastructure, including userspace
interrupt delegation and potentially userspace async runtimes, working before
we tackle this.