CI: work around issues with `rustdoc` generation to fix CI
At some point we'll want to address this proplerly but this should be
sort of fixed for now by just documenting a single top-level crate with
the correct target. In future, we'll want to build each top-level crate
with the correct target and context and then somehow stick all the crates
together at the end.