`kernel`: implement experimental `poll_interest` system call
To support a userspace async runtime, we need to able to tell if something
interesting has happened to a kernel object without actually asking for
that thing (e.g. accepting a message from a channel). To do this initially,
this is an experimental system call to just check if something 'interesting'
has happened to a kernel object, for a fairly arbitrary and object-specific
meaning of interesting.

This doesn't feel like a very considered API yet, hence the experimentalness.
It may well be replaced in the future but allows us to get something working.