`kernel`: better logging of system calls and early userland logging
The Really Bad News: we've started seeing spurious crashes on RISC-V that
look awfully, horrendously similar to the horrible un-debuggable ones we
got on x64 (and still aren't solved). This is possibly the worst news I've
ever received.

The Okay News: in trying to debug it (and failing), I made the `syscall`
logging slightly nicer to look at, and included the task name making the
system call