`usb_hid`: rewrite to use the userland async runtime
This will allow us to support multiple HID devices at the same time,
which is definitely something we'll need to support for the average platform.
Unfortunately, this is the point The Really Awful Bug seems to be rearing
its ugly head, so I guess it's uncovered by increasing complexity of
userspace tasks? I think we'll need some really heavy duty fuzzing+tracing
to find where the hell it's coming from, so let's ignore it for now.