`kernel`: fix bug in buddy allocator
This is (I think) what was causing the crashes on RISC-V around page
table corruption. I feel a bit stupid that we managed to make this mistake
but it is what it is - basically we were constructing higher-order blocks
incorrectly by not checking that they were well-aligned to their larger
sizes. When they were split, this meant buddies that were never part of
allocatable memory blocks were added, so we were double-allocating memory.

I was really hoping that this was the root cause of all of our issues, but
this does NOT seem to fix the crashes on x86_64. Maybe there are two
underlying issues after all...