`usb_bus_ehci`: provide access to `Controller` for device handling tasks
We'll need access to the `Controller` in `ActiveDevice`, which is tricky
from inside `Controller` itself because of lifetimes etc. We solve this by
passing out a list of new devices from `Controller::check_ports` and then
spawning the new tasks from outside the controller. Starting to wonder if
this means more device-level logic should move from `Controller` to `ActiveDevice`
now?