`ptah_derive`: relax restriction on Ptah + descriminated enums
We want to use an enum with descriminators with Ptah. This is potentially
tricky, but can also be worked around easily by ignoring the descriminators.
However, this could create issues in the future if users expect the descriminators
to accurately appear in the wire format (which with this solution they do
not).