`ptah`: use real `ptah_derive` for all dependencies
I didn't really realise that `patch`s in `Cargo.toml`s don't apply to downstream
users of a crate, although it does of course make perfect sense. We want
to use the local version of `ptah_derive` throughout the repo to avoid
confusion, so let's just set it here (and will have to change it temporarily
to publish).