`usb_bus_ehci`: support basic device control from the driver channel
This adds the functionality we need to drive a HID device on QEMU in the
simplest way possible. It's a bit of a jumble of functionality as it has
been done restrospectively but this is an exciting milestone for the USB
stack! Some bits will require cleanup, especially when transactions are
tracked via the async system.