ginkgo: move token data out of `TokenType`
This allows `TokenType` to be used more easily to match tokens in the parser,
and has the added advantage of shrinking tokens by quite a bit. Values
are now extracted via methods on `Lex`, but the ergonomics of this (esp
when the lexer is wrapped in a `PeekingIter`) could be improved in future.