user/virtio_gpu: start a hacky Virtio GPU driver
This drives a Virtio GPU device enough to draw a framebuffer in guest memory
to the display! Various parts are still pretty hacky so we'll need various
bits of infrastructure in the PCI system to clean things up, but we should
be able to get a basic graphical shell working with this!