`platform_bus`: reverse matching of device interest filtering
We chose before to need ALL of the supplied filters to match for a device
to be handed off to a driver. I actually think the reverse would be more
useful: if any of the top-level filters match, the driver is asked if it
can support a device, and filters can be conjugated using an adaptor.
This lets drivers show interest in multiple types of device more easily.