`platform_bus`: move `pci_bus` functionality into Platform Bus itself
It felt kind of unnecessary to have `pci_bus` spin up as a whole task
just to munge a tiny bit of data from the kernel and then spin forever,
and PCI devices are common across many of our supported platforms. I thought
it made more sense to have this functionality live directly inside of the
Platform Bus.