`kernel`: make `wait_for_event` syscall capable of being non-blocking
This is a little unintuitive, so we may want to change it in the future.
We want `Event`s to be usable from an async context, and so need to clear
an event that may be not-pending, but without blocking. We detect if an
event is pending using `poll_interest`, and then we can use `wait_for_event`
to clear it. We don't want this call to block if e.g. the event has been
cancelled, so we'll add a parameter to demark whether we want it to block.