`usb_bus_ehci`: use separate task to handle controller interrupts
This requires us to remove the overall lock on `Controller`, and instead
lock each part of it. The locking strategy is very naive at the moment
as a proof-of-concept, and should probably be improved in the future.

Obviously just having a global `WaitCell` is temporary - this waiter will
be moved into individual queues or maybe onto transactions so devices can
await the thing they're actually interested in.