`kernel`: add support for PCI devices that need legacy interrupt routing
The QEMU EHCI device does not actually have MSI support, we bodged it in
for a while with a custom QEMU fork. This commit adds legacy PCI interrupt
pin support in the common kernel + the RISC-V implementation, with it stubbed
out for now on x86_64. The FDT code is pretty botchy for this, but probably
needs a larger overhaul in the future so will do for now.