`kernel`: move to concept of 'bootstrap' task + others
This is the initial part of the move to having a single 'bootstrap' task
that takes all other tasks and starts them from userspace. This will hopefully
allow us to move service discovery into userspace and in general be a lot
more flexible.

This breaks Poplar as a single commit, as only a single userspace task will
start, but this will be fixed in future commits.