`lib/std`: stack-allocate panic message on userspace panic
This was a slightly frustrating bug to track down, as I'd forgotten we'd
done this like this. Exhausting the heap in userspace produces a stack
overflow, as we attempt to panic, try to allocate a buffer to put the panic
message into, and then panic again as we can't allocate. This repeats until
the stack overflows and you trap...

An initial fix can be done by stack-allocating a buffer for the panic message,
which will work as long as the message isn't too long and we don't overflow
the stack while formatting the message. This shouldn't be a problem for
normal-sized messages I don't think.