`xtask`: add workaround around QEMU's poor mouse capturing
Current versions of QEMU seem to have a bug (at least on my machine) where
a captured mouse cursor on the GTK backend is not generating mouse movement
events on an emulated USB mouse on the guest. Forcing GTK to use its X
backend seems to fix this, so do that for now.