`kernel`: rework context switches to use arch task contexts
Previously, we had a bit of a cackhanded system of tracking both arch-
dependent task contexts, as well as kernel and user stack pointers, in
the common part of the scheduler. Moving x64 to store its stack pointers
in the context has both cleaned up the code across the kernel, and has
actually seemed to fix the slightly flaky context switch bug on x64!