`kernel`: dynamically allocate kernel virtual memory for MMIO
Now that Seed gives us a smaller direct mapping of just conventional memory,
we need a more advanced way of managing virtual memory for MMIO. This
moves to a better layout for the higher-half and introduces an area of
virtual memory that the kernel can dynamically allocate for such uses.

Various parts of the kernel will need to be updated to use this new system.

Also splits the `kernel::memory` module into new `pmm` and `vmm` modules at
the top-level. This feels appropriate as the PMM and VMM are much more separate
concerns that I believed when I first wrote this part of the kernel.