`seed`,`kernel`: get RV Seed recompiling, move to dynamic memory layout approach
This is required for platforms with differing memory layouts depending on
exact configuration, such as RV64. This is a bit tedious - maybe we should
reconsider our support for such an array of platforms for RISC-V? The
actual boot appears to be broken - I wonder if this is a longer-term issue
that we haven't picked up due to lack of testing on the target?